Move Tri/QuadFunction to amidst.utils; some Gson stuff
