+++ src/main/java/amidst/mojangapi/world/versionfeatures/VersionFeatures.java
@@ -9,6 +9,8 @@ import amidst.mojangapi.world.icon.locationchecker.LocationChecker;
 import amidst.mojangapi.world.icon.locationchecker.MineshaftAlgorithm_Base;
 import amidst.mojangapi.world.icon.producer.StrongholdProducer_Base;
 import amidst.mojangapi.world.oracle.BiomeDataOracle;
+import amidst.util.QuadFunction;
+import amidst.util.TriFunction;
 
 @Immutable
 public class VersionFeatures {
